Updated twigs to use 32 bit ints

+++ src/Debug.cpp
@@ -102,7 +102,7 @@ static void printTree(const Ocroot *r, uint32_t t)
     Octype tp = (Octype)r->tree[t].type();
     printf("%08x:[%s:%08x]\n", t, tps[tp], (uint32_t)r->tree[t].offset());
     if (tp == TWIG)
-        printf("\t%016llx\n", r->twig[r->tree[t].offset()].leafmap[0]);
+        printf("\t%016llx\n", (uint64_t)r->twig[r->tree[t].offset()].leafmap[0]);
     if (tp == BRANCH)
     {
         for (int i = 0; i < 8; ++i)
